Added measurement to moveAlongPath and found right condition for communication event. Need to fix updatePaths function so that I can just update paths for a single team of robots and not for all teams in sched(k)

+++ src/Classes/Robot.py
@@ -94,7 +94,7 @@ class Robot:
         # newData = new measurement for single robot
         # currentLocations = sensing location of single robot
         
-        self.mapping[currentLocations[0],currentLocations[1]] = newData
+        self.mapping[np.int(currentLocations[0]),np.int(currentLocations[1])] = newData
     
     @classmethod     
     def getTotalMap(cls):
